+++ gcs/src/main/java/com/google/cloud/hadoop/fs/gcs/GoogleHadoopFileSystemConfiguration.java
@@ -16,6 +16,7 @@
 
 package com.google.cloud.hadoop.fs.gcs;
 
+import static com.google.cloud.hadoop.gcsio.GoogleCloudStorageOptions.AUTO_REPAIR_IMPLICIT_DIRECTORIES_DEFAULT;
 import static com.google.cloud.hadoop.util.HadoopCredentialConfiguration.HTTP_TRANSPORT_SUFFIX;
 import static com.google.cloud.hadoop.util.HadoopCredentialConfiguration.PROXY_ADDRESS_SUFFIX;
 import static com.google.cloud.hadoop.util.HadoopCredentialConfiguration.PROXY_PASSWORD_SUFFIX;
@@ -115,7 +116,9 @@ public class GoogleHadoopFileSystemConfiguration {
    * authority corresponding to that root path in the process.
    */
   public static final HadoopConfigurationProperty<Boolean> GCE_BUCKET_DELETE_ENABLE =
-      new HadoopConfigurationProperty<>("fs.gs.bucket.delete.enable", false);
+      new HadoopConfigurationProperty<>(
+          "fs.gs.bucket.delete.enable",
+          GoogleCloudStorageFileSystemOptions.DEFAULT.isBucketDeleteEnabled());
 
   /** Configuration key for GCS project ID. Default value: "DISABLED" */
   public static final HadoopConfigurationProperty<RequesterPaysMode> GCS_REQUESTER_PAYS_MODE =
@@ -144,7 +147,9 @@ public class GoogleHadoopFileSystemConfiguration {
    * conjunction with other caching options.
    */
   public static final HadoopConfigurationProperty<Boolean> GCS_PERFORMANCE_CACHE_ENABLE =
-      new HadoopConfigurationProperty<>("fs.gs.performance.cache.enable", false);
+      new HadoopConfigurationProperty<>(
+          "fs.gs.performance.cache.enable",
+          GoogleCloudStorageFileSystemOptions.DEFAULT.isPerformanceCacheEnabled());
 
   /**
    * Configuration key for maximum number of milliseconds a GoogleCloudStorageItemInfo will remain
@@ -160,7 +165,9 @@ public class GoogleHadoopFileSystemConfiguration {
    * parallel to reduce latency.
    */
   public static final HadoopConfigurationProperty<Boolean> GCS_STATUS_PARALLEL_ENABLE =
-      new HadoopConfigurationProperty<>("fs.gs.status.parallel.enable", true);
+      new HadoopConfigurationProperty<>(
+          "fs.gs.status.parallel.enable",
+          GoogleCloudStorageFileSystemOptions.DEFAULT.isStatusParallelEnabled());
 
   /** Configuration key for enabling lazy initialization of GCS FS instance. */
   public static final HadoopConfigurationProperty<Boolean> GCS_LAZY_INITIALIZATION_ENABLE =
@@ -171,7 +178,17 @@ public class GoogleHadoopFileSystemConfiguration {
    * inside delete and rename calls.
    */
   public static final HadoopConfigurationProperty<Boolean> GCS_REPAIR_IMPLICIT_DIRECTORIES_ENABLE =
-      new HadoopConfigurationProperty<>("fs.gs.implicit.dir.repair.enable", true);
+      new HadoopConfigurationProperty<>(
+          "fs.gs.implicit.dir.repair.enable", AUTO_REPAIR_IMPLICIT_DIRECTORIES_DEFAULT);
+
+  /**
+   * Configuration key for enabling check to ensure that conflicting directories do not exist when
+   * creating files and conflicting files do not exist when creating directories.
+   */
+  public static final HadoopConfigurationProperty<Boolean> GCS_CREATE_ITEMS_CONFLICT_CHECK_ENABLE =
+      new HadoopConfigurationProperty<>(
+          "fs.gs.create.items.conflict.check.enable",
+          GoogleCloudStorageFileSystemOptions.DEFAULT.isEnsureNoConflictingItems());
 
   /** Configuration key for customizing glob search algorithm. */
   public static final HadoopConfigurationProperty<GlobAlgorithm> GCS_GLOB_ALGORITHM =
@@ -363,7 +380,9 @@ public class GoogleHadoopFileSystemConfiguration {
    * isolation.
    */
   public static final HadoopConfigurationProperty<Boolean> GCS_COOPERATIVE_LOCKING_ENABLE =
-      new HadoopConfigurationProperty<>("fs.gs.cooperative.locking.enable", false);
+      new HadoopConfigurationProperty<>(
+          "fs.gs.cooperative.locking.enable",
+          GoogleCloudStorageFileSystemOptions.DEFAULT.isCooperativeLockingEnabled());
 
   /** Configuration key for lock expiration when using cooperative locking. */
   public static final HadoopConfigurationProperty<Long>
@@ -411,14 +430,16 @@ public class GoogleHadoopFileSystemConfiguration {
   // @VisibleForTesting
   static GoogleCloudStorageFileSystemOptions.Builder getGcsFsOptionsBuilder(Configuration config) {
     return GoogleCloudStorageFileSystemOptions.builder()
+        .setCloudStorageOptions(getGcsOptionsBuilder(config).build())
         .setBucketDeleteEnabled(GCE_BUCKET_DELETE_ENABLE.get(config, config::getBoolean))
-        .setMarkerFilePattern(GCS_MARKER_FILE_PATTERN.get(config, config::get))
-        .setPerformanceCacheEnabled(GCS_PERFORMANCE_CACHE_ENABLE.get(config, config::getBoolean))
         .setCooperativeLockingEnabled(
             GCS_COOPERATIVE_LOCKING_ENABLE.get(config, config::getBoolean))
+        .setEnsureNoConflictingItems(
+            GCS_CREATE_ITEMS_CONFLICT_CHECK_ENABLE.get(config, config::getBoolean))
+        .setMarkerFilePattern(GCS_MARKER_FILE_PATTERN.get(config, config::get))
+        .setPerformanceCacheEnabled(GCS_PERFORMANCE_CACHE_ENABLE.get(config, config::getBoolean))
         .setPerformanceCacheOptions(getPerformanceCachingOptions(config))
-        .setStatusParallelEnabled(GCS_STATUS_PARALLEL_ENABLE.get(config, config::getBoolean))
-        .setCloudStorageOptions(getGcsOptionsBuilder(config).build());
+        .setStatusParallelEnabled(GCS_STATUS_PARALLEL_ENABLE.get(config, config::getBoolean));
   }
 
   @VisibleForTesting