+++ gcsio/src/main/java/com/google/cloud/hadoop/gcsio/GoogleCloudStorageFileSystem.java
@@ -231,8 +231,9 @@ public class GoogleCloudStorageFileSystem {
               "Cannot create a file whose name looks like a directory: '%s'", resourceId));
     }
 
-    // Before creating a leaf directory we need to check if there are no conflicting files
-    // with the same name as any subdirectory
+    // Before creating a file we need to check:
+    // - if there are no conflicting files with the same name as any parent subdirectory
+    // - if there are no conflicting directory with the name as a file
     if (options.isEnsureNoConflictingItems()) {
       // Asynchronously check if a directory with the same name exists.
       StorageResourceId dirId = resourceId.toDirectoryId();
@@ -244,7 +245,7 @@ public class GoogleCloudStorageFileSystem {
 
       checkNoFilesConflictingWithDirs(resourceId);
 
-      // Check if a directory of that name exists.
+      // Check if a directory with the same name exists.
       if (getFromFuture(conflictingDirExist)) {
         throw new FileAlreadyExistsException("A directory with that name exists: " + path);
       }